Remove unnecessary require; move resource-label into resource to avoid circular dep.; document
