--- app/common/components/resource-label/label/__init__.js
@@ -1,3 +1,0 @@
-module.exports = {
-  PATH: require('./../__init__').PATH + '/label'
-}
