+++ app/common/components/resource/resource-label/label-button/label-button.directive.html
