+++ app/common/components/resource/resource-label/label-edit/label-edit.directive.sass
