+++ app/common/components/resource/resource-label/label-modal.config.js
