+++ app/common/components/resource/resource-label/label.controller.html
