+++ app/common/components/resource/resource-label/label.controller.sass
@@ -1,4 +1,4 @@
-@import '../../styles/variables'
+@import '../../../styles/variables'
 #labels > *
   display: inline-block
 #labels .resource-label
