+++ app/common/components/resource/resource-label/label/label.directive.js
