+++ app/common/components/resource/resource-label/label/label.directive.sass
@@ -1,4 +1,4 @@
-@import '../../../styles/variables'
+@import '../../../../styles/variables'
 
 .resource-label
   width: 20em
