+++ app/common/components/resource/resource-label/labels-to-pdf.js
